+++ backend/src/models/projects/journal_voucher.js
@@ -1,6 +1,5 @@
 const dbConnection = require("@database/databaseConnection");
 const { knex, dataBaseSchemas } = dbConnection();
-const { dateFormat } = require("../../helpers/standards");
 const jvTable = `${dataBaseSchemas().data}.jv`;
 const fiscalYearTable = `${dataBaseSchemas().data}.fiscal_year`;
 const clientCodingTable = `${dataBaseSchemas().data}.client_coding`;
@@ -11,7 +10,7 @@ const findAll = (projectId) => {
     .columns(
       "jv.id",
       "jv.jv_number",
-      { billed_date: knex.raw(`TO_CHAR(jv.billed_date :: DATE, '${dateFormat}')`) },
+      "jv.billed_date",
       "jv.amount",
       "jv.quarter",
       "jv.project_id",