+++ frontend/src/pages/Contracts/Contract/Resources/tableConfig.tsx
@@ -1,4 +1,5 @@
 import { GridColDef } from "@mui/x-data-grid";
+import { dateFormatter as valueFormatter } from "utils/formatDate";
 
 export const tableConfig = () => {
   const defaultFlex = 3;
@@ -11,8 +12,8 @@ export const tableConfig = () => {
     { field: "assignment_rate", headerName: "Assignment Rate", flex: defaultFlex },
     { field: "hours", headerName: "Hours", flex: defaultFlex },
     { field: "fees_for_resource", headerName: "Fees For Resource", flex: defaultFlex },
-    { field: "start_date", headerName: "Start Date", flex: defaultFlex },
-    { field: "end_date", headerName: "End Date", flex: defaultFlex },
+    { field: "start_date", headerName: "Start Date", valueFormatter, flex: defaultFlex },
+    { field: "end_date", headerName: "End Date", valueFormatter, flex: defaultFlex },
   ];
 
   const initialState = {